+++ packages/react/src/index.ts
@@ -3,13 +3,11 @@ import Block from './Block';
 import curious from './curious';
 import Link from './Link';
 import Navigator from './Navigator';
-import Provider from './Provider';
 
 export {
   Active,
   Block,
   curious,
   Link,
-  Navigator,
-  Provider
+  Navigator
 };