+++ packages/react/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Remove `<Provider>`. It was only adding half of the necessary `context` values (did not provide the `response`). The `<Navigator>` should be used instead.
+
 ## 1.0.0-beta.9
 
 * Remove `<Redirect>`. This was inspired by React Router, but it doesn't really serve a point. Redirects should either be known pre-render, in which case they can be made in `route.load` or can be manually performed using `curi.history.replace`.